[NUB-1034] Model float filter

+++ packages/core/src/components/forms/fields/FieldFloat.js
@@ -2,7 +2,7 @@ import FieldInput from './FieldInput';
 
 // We want it to be blank if it's not a number
 const setValueAs = (value) => {
-  const parsed = parseInt(value, 10);
+  const parsed = parseFloat(value, 10);
   if (isNaN(parsed)) return undefined;
 
   return parsed;
